requirements.txt
# Gmail API dependencies
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Google ADK dependencies (if not already included)
google-adk

# Additional dependencies for the project
requests==2.31.0
Pillow==10.1.0
google-generativeai==0.3.2

# Research query dependencies
dedalus-labs==0.1.0a8
httpx==0.27.0

# Development dependencies (optional)
python-dotenv==1.0.0
